% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mimu.R
\name{make_mimu}
\alias{make_mimu}
\title{Creates a Multivariate IMU object.}
\usage{
make_mimu(..., freq, unit = NULL, sensor.name = NULL, exp.name = NULL,
  for_test = NULL)
}
\arguments{
\item{...}{A \code{list} of \code{vector} reprenting multiple error signals from IMU}

\item{freq}{A \code{numeric} value with frequency of IMU.}

\item{sensor.name}{A \code{string} with name of IMU.}

\item{exp.name}{A \code{vector} with the experiment name}

\item{units}{A \code{string} value with unit of IMU.}
}
\value{
A \code{mimu} object with the structure:
\describe{
\item{data}{original error signal for imu}
\item{variance}{Data empirical Wavelet Variance}
\item{ci_low}{Lower Confidence Interval for Wavelet Variance}
\item{ci_high}{Upper Confidence Intervalfor Wavelet Variance}
\item{scales}{Scales of the Wavelet Variance}}
}
\description{
This function creates a multivariate IMU or MIMU object, combining multiple replicates of the
same inertial sensor.
}
\examples{
n = 10^6
Xt = rnorm(n/4)
Yt = rnorm(n/2) + cumsum(rnorm(n/2, 0, 10^(-2)))
Zt = rnorm(n) + cumsum(rnorm(n, 0, 10^(-3)))
obj = make_mimu(Xt, Yt, Zt, freq = 100, unit = "s",
sensor.name = "MTiG - Gyro. X", exp.name = c("today", "yesterday", "a few days ago"))
}
\author{
Gaetan Bakalli and Stephane Guerrier
}
