% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mimu_obj.R
\name{make_wvar_mimu_obj}
\alias{make_wvar_mimu_obj}
\title{Creates a Multivariate IMU object.}
\usage{
make_wvar_mimu_obj(..., freq, unit = NULL, sensor.name = NULL,
  exp.name = NULL, for_test = NULL)
}
\arguments{
\item{...}{A \code{list} of \code{vector} reprenting multiple error signals from IMU}

\item{freq}{A \code{numeric} value with frequency of IMU.}

\item{sensor.name}{A \code{string} with name of IMU.}

\item{exp.name}{A \code{vector} with the nemane}

\item{units}{A \code{string} value with unit of IMU.}
}
\description{
This function creates a Multivariate IMU objects, combining multiple replicates of the
same inertial sensor.
}
\examples{
n = 10^6
Xt = rnorm(n/4)
Yt = rnorm(n/2) + cumsum(rnorm(n/2, 0, 10^(-2)))
Zt = rnorm(n) + cumsum(rnorm(n, 0, 10^(-3)))
obj = make_wvar_mimu_obj(Xt, Yt, Zt, freq = 100, unit = "s",
sensor.name = "MTiG - Gyro. X", exp.name = c("today", "yesterday", "a few days ago"))
}
\author{
Stephane Guerrier
}
