% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mgmwm_obj_function}
\alias{mgmwm_obj_function}
\title{Objective function Multivariate GMWM}
\usage{
mgmwm_obj_function(theta, model, mimu)
}
\arguments{
\item{theta}{A \code{vector} of parameter from a model.}

\item{model}{A model \code{object} of class \code{ts.model}.}

\item{mimu}{A mimu \code{object}.}
}
\value{
\code{numeric} value of the objective function
}
\description{
Compute the objective function Multivariate GMWM, based on an \code{mimu} object and a \code{model}
}
\author{
Gaetan Bakalli
}
\keyword{internal}
