% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgmwm.R
\name{plot.mgmwm}
\alias{plot.mgmwm}
\title{plot method for mgmwm}
\usage{
\method{plot}{mgmwm}(
  object,
  decomp = FALSE,
  add_legend_mgwmw = TRUE,
  legend_pos = NULL,
  ylab_mgmwm = NULL
)
}
\description{
plot method for mgmwm
}
