% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mimu.R, R/utils.R
\name{wv_theo}
\alias{wv_theo}
\alias{wv_theo}
\title{Compute theroretical Wavelet Variance}
\usage{
\method{plot}{mimu}(model, tau)

wv_theo(model, tau)
}
\arguments{
\item{model}{A \code{model} object.}

\item{tau}{A \code{vector} of \eqn{2^1:number of scales}}

\item{model}{A \code{model} object.}

\item{tau}{A \code{vector} of \eqn{2^1:number of scales}}
}
\value{
\code{vector} of theoretical Wavelet Variance for a given model

\code{vector} of theoretical Wavelet Variance for a given model
}
\description{
Displays a plot of wavelet variances accounting for CI values.

Compute theroretical Wavelet Variance
}
\examples{
#' Xt = rnorm(n/4)
Yt = rnorm(n/2) + cumsum(rnorm(n/2, 0, 10^(-2)))
Zt = rnorm(n) + cumsum(rnorm(n, 0, 10^(-3)))
obj = make_mimu(Xt, Yt, Zt, freq = 100, unit = "s",
sensor.name = "MTiG - Gyro. X", exp.name = c("today", "yesterday", "a few days ago"))
model = AR1() + WN()

Extract the scales of first replicates of mimu \\code{object} and compute tau
tau = mimu[[1]]$tau
wv_theo = function(model, tau)
#' Xt = rnorm(n/4)
Yt = rnorm(n/2) + cumsum(rnorm(n/2, 0, 10^(-2)))
Zt = rnorm(n) + cumsum(rnorm(n, 0, 10^(-3)))
obj = make_wvar_mimu_obj(Xt, Yt, Zt, freq = 100, unit = "s",
sensor.name = "MTiG - Gyro. X", exp.name = c("today", "yesterday", "a few days ago"))
model = AR1() + WN()

Extract the scales of first replicates of mimu \\code{object} and compute tau
tau = 2^(1:length(sensor.name[[1]]$scales))
wv_theo = function(model, tau)
}
\author{
Stephane Guerrier, Nathanael Claussen, and Justin Lee

Stephane Guerrier, Nathanael Claussen, and Justin Lee
}
\keyword{internal}
